Push

+++ src/horst/IBewegung.java
@@ -6,7 +6,7 @@ public interface IBewegung {
 	
 	public boolean turn (double degree);
 	
-	public boolean move (int distance);
+	public boolean move (double distance);
 	
 	public boolean goRichtung (Richtung richtung);
